Fixed the simple import example.
In first import example, imports happen in 3 steps:

- create empty dataset
- upload a source
- add a batch to the dataset from the source

For upload a source, the example showed posting a CSV body. This
doesn't work. You can post a json body, but CSV has to be part of an
multi-part form upload.  I've updated the example to reflect that.

Perhaps a CSV body should be accepted. I'm going to be in that code
to add zip file support, so I can allow zip/csv bodies as well.

In the 3rd step, the example left out the element property, which is
required. The error message:

The payload of a POST to the batches catalog must be either:
    * A file upload, with a Content-Type from the set ['text/csv', 'application/csv', 'application/x-spss-sav'],
    * A multipart/form-data upload, with a "file" parameter
        whose Content-Type is from the above set,
    * A crunch:table containing data and, optionally, metadata, or
    * A shoji:entity representing the new Batch. Its "body" must contain either:
        * A 'source' member, with the URL of an existing Source,
        * A 'url' member, with the URL of an s3 resource,
        * A 'dataset' member, with the URL of a different dataset, or
        * A 'stream' member, with the number of
            streamed messages to consume (null for all);
            you must also include a 'type' member defining
            the file type of the streamed rows ('ldjson' or 'csv').

It looks like the 2nd and 3rd steps could be combined. Perhaps that
would be better.

I wonder what percentage of data sets have more than one batch. If the
common case is a single batch, it seems a bit mean to make people go
through multiple steps. <shrug>