Added a note that a zip file may be used.
Note that I only updated the "Import from a data file" section.

I didn't touch the "Metadata document + CSV" sections because, as the
section is now written, uploading a Zip file won't work.  This is
because posting content directly checks that the content type is CSV
or SPSS, and I didn't include changing that code in this PR.  OTOH,
Zipe import does work in a multi-part post.